Consolidate Diagram
