+++ src/App.tsx
@@ -18,13 +18,13 @@ import { BitcoinNodeManager, update_broadcastable } from './Data/BitcoinNode';
 import { ContractModel, Data, timing_cache } from './Data/ContractManager';
 import { TransactionModel, PhantomTransactionModel } from './Data/Transaction';
 import { UTXOModel } from './Data/UTXO';
-import { SpendPortModel } from './DiagramComponents/SpendLink/SpendLink';
-import { SpendLinkFactory } from './DiagramComponents/SpendLink/SpendLinkFactory';
-import { TransactionNodeFactory } from './DiagramComponents/TransactionNode/TransactionNodeFactory';
-import { UTXONodeFactory } from './DiagramComponents/UTXONode/UTXONodeFactory';
+import { SpendPortModel } from './UX/Diagram/DiagramComponents/SpendLink/SpendLink';
+import { SpendLinkFactory } from './UX/Diagram/DiagramComponents/SpendLink/SpendLinkFactory';
+import { TransactionNodeFactory } from './UX/Diagram/DiagramComponents/TransactionNode/TransactionNodeFactory';
+import { UTXONodeFactory } from './UX/Diagram/DiagramComponents/UTXONode/UTXONodeFactory';
 import { SimulationController } from './Simulation';
 import { AppNavbar } from './UX/AppNavbar';
-import { DemoCanvasWidget } from './UX/DemoCanvasWidget';
+import { DemoCanvasWidget } from './UX/Diagram/DemoCanvasWidget';
 import { EmptyViewer, EntityViewerModal, Viewer } from './UX/EntityViewer';
 import Collapse from 'react-bootstrap/Collapse';
 import './Glyphs.css';
