+++ src/UX/Diagram/DemoCanvasWidget.tsx
