+++ src/UX/Diagram/DiagramComponents/OutputLink.js
