+++ src/UX/Diagram/DiagramComponents/SpendLink/SpendLink.js
