+++ src/UX/Diagram/DiagramComponents/SpendLink/SpendLinkFactory.tsx
