+++ src/UX/Diagram/DiagramComponents/SpendLink/SpendLinkModel.ts
