+++ src/UX/Diagram/DiagramComponents/SpendLink/SpendLinkWidget.tsx
