+++ src/UX/Diagram/DiagramComponents/TransactionNode/Ants.css
