+++ src/UX/Diagram/DiagramComponents/TransactionNode/TransactionNodeModel.js
