+++ src/UX/Diagram/DiagramComponents/TransactionNode/TransactionNodeWidget.tsx
