+++ src/UX/Diagram/DiagramComponents/UTXONode/Ants.css
